Rename Duration namespace from Eophantasy\Type\Time\Duration to Eophantasy\Time\Duration
